+++ docs/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 * Collection exotics are no longer duplicated. They are also sorted by name.
 * Updated max power to 380.
+* Vendors and collections will no longer show items exclusive to platforms other than the current account's platform.
 
 # 4.51.2 (2018-05-09)